fix mistakes

+++ src/services/contacts.js
@@ -1,20 +1,20 @@
 import { Contact } from '../db/models/contact.js'; // ✅ виправлений імпорт
 
 export const getContacts = (userId) => {
-  return Contact.find({ userId });
+  return Contact.find({ owner: userId });
 };
 
 export const getContactById = (id, userId) => {
-  return Contact.findOne({ _id: id, userId });
+  return Contact.findOne({ _id: id, owner: userId });
 };
 
 export const createContact = (data, userId) => {
-  return Contact.create({ ...data, userId });
+  return Contact.create({ ...data, owner: userId });
 };
 
 export const updateContact = (id, data, options = {}, userId) => {
   return Contact.findOneAndUpdate(
-    { _id: id, userId },
+    { _id: id, owner: userId },
     data,
     { ...options, new: true }
   ).then((contact) => ({
@@ -24,5 +24,6 @@ export const updateContact = (id, data, options = {}, userId) => {
 };
 
 export const deleteContact = (id, userId) => {
-  return Contact.findOneAndDelete({ _id: id, userId });
+  return Contact.findOneAndDelete({ _id: id, owner: userId });
 };
+
